+++ scripts/changelog.js
@@ -62,7 +62,7 @@ Hooks.once("ready", () => {
       version: "0.1.2",
       title: "NPC Leader Option",
       body: `\
-          Per @GhostwheelX's request, added a game setting that to choose the NPC leader based on the best NPC initiative roll,
+          Per @GhostwheelX's request, I added a game setting that to choose the NPC leader based on the best NPC initiative roll,
           instead of by best NPC bonus to initiative. Note that you can still manually roll a single NPC, which will then become the leader,
           or multiple NPCs, in which case the leader will be chosen amongst those manually rolled.
           `
@@ -72,9 +72,25 @@ Hooks.once("ready", () => {
       version: "0.1.3",
       title: "Dice-So-Nice Setting",
       body: `\
-          Per @noncasus's request, added a game setting that will force Dice-So-Nice initiative rolls when rolling for
+          Per @noncasus's request, I added a game setting that will force Dice-So-Nice initiative rolls when rolling for
           all NPCs or all tokens. The rolls reflect real calculations that are used to order the NPCs, but
           ultimately their initiative may differ dramatically from the rolls to facilitate the zipper initiative order.
+          `
+    })
+
+    .addEntry({
+      version: "0.2.1",
+      title: "Zipper Popcorn and Interleaving",
+      body: `\
+          Per @EpsilonRose's request, I added two settings that enable a "zipper popcorn" option for PC and NPC groups.
+          When a PC ends their turn with 2+ PCs left in the initiative order, a dialog will allow the PC to
+          select the next PC to go (still in zipper order). Essentially, the next PC in the initiative and the
+          chosen PC switch spots. A similar dialog is presented to the GM when 1+ PCs remain in the initiative.
+
+          Per @GhostwheelX's request, I added a setting to interleave the NPCs when there are more NPCs than PCs.
+          Spacing will be approximately equal. For example, with 2 PCs and 5 NPCs, you could get
+          P N N P N N instead of
+          P N P N N N.
           `
     })