+++ scripts/settings.js
@@ -4,10 +4,10 @@ game
 
 "use strict";
 
-import { MODULE_ID } from "./const.js";
+import { ModuleSettingsAbstract } from "./ModuleSettingsAbstract.js";
 import { registerPopcorn, deregisterPopcorn } from "./patching.js";
 
-export const SETTINGS = {
+const SETTINGS = {
   CHANGELOG: "changelog",
   RESET_EACH_ROUND: "resetEachRound",
   NPC_LEADER_HIGHEST_INIT: "npcLeaderHighestInitiative",
@@ -19,83 +19,76 @@ export const SETTINGS = {
   }
 };
 
-export function getSetting(settingName) {
-  return game.settings.get(MODULE_ID, settingName);
-}
+export class Settings extends ModuleSettingsAbstract {
+  /** @type {object} */
+  static KEYS = SETTINGS;
 
-export async function toggleSetting(settingName) {
-  const curr = getSetting(settingName);
-  return await game.settings.set(MODULE_ID, settingName, !curr);
-}
+  static registerAll() {
+    const { KEYS, register, localize } = this;
 
-export async function setSetting(settingName, value) {
-  return await game.settings.set(MODULE_ID, settingName, value);
-}
+    register(KEYS.RESET_EACH_ROUND, {
+      name: localize(`${KEYS.RESET_EACH_ROUND}.Name`),
+      hint: localize(`${KEYS.RESET_EACH_ROUND}.Hint`),
+      type: Boolean,
+      default: false,
+      scope: "world",
+      config: true
+    });
 
-export function registerSettings() {
-  game.settings.register(MODULE_ID, SETTINGS.RESET_EACH_ROUND, {
-    name: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.RESET_EACH_ROUND}.Name`),
-    hint: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.RESET_EACH_ROUND}.Hint`),
-    type: Boolean,
-    default: false,
-    scope: "world",
-    config: true
-  });
+    register(KEYS.NPC_LEADER_HIGHEST_INIT, {
+      name: localize(`${KEYS.NPC_LEADER_HIGHEST_INIT}.Name`),
+      hint: localize(`${KEYS.NPC_LEADER_HIGHEST_INIT}.Hint`),
+      type: Boolean,
+      default: false,
+      scope: "world",
+      config: true
+    });
 
-  game.settings.register(MODULE_ID, SETTINGS.NPC_LEADER_HIGHEST_INIT, {
-    name: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.NPC_LEADER_HIGHEST_INIT}.Name`),
-    hint: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.NPC_LEADER_HIGHEST_INIT}.Hint`),
-    type: Boolean,
-    default: false,
-    scope: "world",
-    config: true
-  });
+    register(KEYS.USE_DSN, {
+      name: localize(`${KEYS.USE_DSN}.Name`),
+      hint: localize(`${KEYS.USE_DSN}.Hint`),
+      type: Boolean,
+      default: true,
+      scope: "world",
+      config: true
+    });
 
-  game.settings.register(MODULE_ID, SETTINGS.USE_DSN, {
-    name: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.USE_DSN}.Name`),
-    hint: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.USE_DSN}.Hint`),
-    type: Boolean,
-    default: true,
-    scope: "world",
-    config: true
-  });
+    register(KEYS.INTERLEAVE_NPCS, {
+      name: localize(`${KEYS.INTERLEAVE_NPCS}.Name`),
+      hint: localize(`${KEYS.INTERLEAVE_NPCS}.Hint`),
+      type: Boolean,
+      default: false,
+      scope: "world",
+      config: true
+    });
 
-  game.settings.register(MODULE_ID, SETTINGS.INTERLEAVE_NPCS, {
-    name: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.INTERLEAVE_NPCS}.Name`),
-    hint: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.INTERLEAVE_NPCS}.Hint`),
-    type: Boolean,
-    default: false,
-    scope: "world",
-    config: true
-  });
+    register(KEYS.POPCORN.PC, {
+      name: localize(`${KEYS.POPCORN.PC}.Name`),
+      hint: localize(`${KEYS.POPCORN.PC}.Hint`),
+      type: Boolean,
+      default: false,
+      scope: "world",
+      config: true,
+      onChange: value => {
+        if ( value ) registerPopcorn();
+        else if ( !this.get(KEYS.POPCORN.NPC) ) deregisterPopcorn();
+      }
+    });
 
-  game.settings.register(MODULE_ID, SETTINGS.POPCORN.PC, {
-    name: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.POPCORN.PC}.Name`),
-    hint: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.POPCORN.PC}.Hint`),
-    type: Boolean,
-    default: false,
-    scope: "world",
-    config: true,
-    onChange: value => {
-      if ( value ) registerPopcorn();
-      else if ( !getSetting(SETTINGS.POPCORN.NPC) ) deregisterPopcorn();
-    }
-  });
+    register(KEYS.POPCORN.NPC, {
+      name: localize(`${KEYS.POPCORN.NPC}.Name`),
+      hint: localize(`${KEYS.POPCORN.NPC}.Hint`),
+      type: Boolean,
+      default: false,
+      scope: "world",
+      config: true,
+      onChange: value => {
+        if ( value ) registerPopcorn();
+        else if ( !this.get(KEYS.POPCORN.PC) ) deregisterPopcorn();
+      }
+    });
 
-  game.settings.register(MODULE_ID, SETTINGS.POPCORN.NPC, {
-    name: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.POPCORN.NPC}.Name`),
-    hint: game.i18n.localize(`${MODULE_ID}.settings.${SETTINGS.POPCORN.NPC}.Hint`),
-    type: Boolean,
-    default: false,
-    scope: "world",
-    config: true,
-    onChange: value => {
-      if ( value ) registerPopcorn();
-      else if ( !getSetting(SETTINGS.POPCORN.PC) ) deregisterPopcorn();
-    }
-  });
-
-
-  if ( getSetting(SETTINGS.POPCORN.PC) || getSetting(SETTINGS.POPCORN.NPC) ) registerPopcorn();
-
+    // Set up the hook to use popcorn.
+    if ( this.get(KEYS.POPCORN.PC) || this.get(KEYS.POPCORN.NPC) ) registerPopcorn();
+  }
 }